requirements.txt
selenium==4.18.1
PySide6==6.8.2
Pillow==10.4.0
reportlab==4.1.0
webdriver-manager==4.0.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3

